Refactored prescriptors to be more user-oriented vs. train oriented. Still have to test them
